Follow changes in rust master and fix bug
* fields are now private by default
* get_xxx operations should not fail if field has no default value

+++ src/lib/core.rs
@@ -195,7 +195,7 @@ impl<'a> CodedInputStream<'a> {
             self.refill_buffer_really();
         }
         assert!(self.buffer_pos < self.buffer_size);
-        let r = self.buffer[self.buffer_pos];
+        let r = self.buffer[self.buffer_pos as uint];
         self.buffer_pos += 1;
         r
     }
@@ -222,7 +222,7 @@ impl<'a> CodedInputStream<'a> {
 
     pub fn read_raw_little_endian32(&mut self) -> u32 {
         let mut bytes = [0u32, ..4];
-        for i in range(0, 4) {
+        for i in range(0u, 4) {
             bytes[i] = self.read_raw_byte() as u32;
         }
         (bytes[0]      ) |
@@ -233,7 +233,7 @@ impl<'a> CodedInputStream<'a> {
 
     pub fn read_raw_little_endian64(&mut self) -> u64 {
         let mut bytes = [0u64, ..8];
-        for i in range(0, 8) {
+        for i in range(0u, 8) {
             bytes[i] = self.read_raw_byte() as u64;
         }
         (bytes[0]      ) |
@@ -453,7 +453,7 @@ impl<'a> CodedOutputStream<'a> {
         if self.position as uint == self.buffer.len() {
             self.refresh_buffer()
         }
-        self.buffer[self.position] = byte;
+        self.buffer[self.position as uint] = byte;
         self.position += 1;
     }
